[Refactor] 프로필 모달에서 로그인 모달 열고 닫기
