+++ src/components/common/Header/HeaderProfileButton.tsx
@@ -5,14 +5,17 @@ import { useDispatch, useSelector } from 'react-redux'
 import HeaderProfileModal from '@/components/common/user/HeaderProfileModal'
 import { RootState } from '@/redux/store'
 import { setModal } from '@/redux/features/profileModalSlice'
+import Portal from '@/portal/Portal'
+import LoginModal from '../user/LoginModal'
 
 function HeaderProfileButton() {
   const dispatch = useDispatch()
 
   const modal = useSelector((state: RootState) => state.profile.modal)
+  const isOpen = useSelector((state: RootState) => state.profile.isOpen)
 
   const handleButtonClick = () => {
-    dispatch(setModal(!modal))
+    dispatch(setModal())
   }
 
   return (
@@ -26,16 +29,15 @@ function HeaderProfileButton() {
             <div className='h-[0px] w-3.5 border border-solid border-gray-800'></div>
             <div className='h-[0px] w-3.5 border border-solid border-gray-800'></div>
           </div>
-          <Image
-            alt={'HeaderProfile'}
-            src={'/images/HeaderProfile.svg'}
-            width={34}
-            height={36}
-            style={{ width: 34, height: 36 }}
-          />
+          <Image alt={'HeaderProfile'} src={'/images/HeaderProfile.svg'} width={34} height={36} />
         </div>
       </button>
       {modal && <HeaderProfileModal />}
+      {isOpen && ( // isOpen을 HeaderProfileModal 안에서 띄우기로 해놨었는데 그럼 종속성 때문에 안됨 그러니까 바깥에 빼야함!!!!!1
+        <Portal>
+          <LoginModal />
+        </Portal>
+      )}
     </div>
   )
 }
