+++ src/components/common/user/HeaderProfileModal.tsx
@@ -1,12 +1,7 @@
 'use client'
 
 import { useState } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
 import HeaderProfileModalItem from '@/components/common/Header/HeaderProfileItem'
-import { setIsOpen } from '@/redux/features/profileModalSlice'
-import { RootState } from '@/redux/store'
-import Portal from '@/portal/Portal'
-import LoginModal from './LoginModal'
 
 const HeaderProfileInfo = [
   {
@@ -32,14 +27,6 @@ const HeaderProfileInfo = [
 ]
 function HeaderProfileModal() {
   const [chosenItem, setChosenItem] = useState('signup')
-  const dispatch = useDispatch()
-  const isOpen = useSelector((state: RootState) => state.profile.isOpen)
-
-  function closeModal() {
-    if (isOpen) {
-      dispatch(setIsOpen(false))
-    }
-  }
 
   return (
     <div className='absolute right-0 top-[65px] z-10 h-[250px] w-[250px] rounded-2xl bg-white shadow-lg'>
@@ -53,11 +40,6 @@ function HeaderProfileModal() {
           />
         ))}
       </ul>
-      {isOpen && (
-        <Portal>
-          <LoginModal onClose={closeModal} />
-        </Portal>
-      )}
     </div>
   )
 }
