+++ src/components/common/user/LoginModal.tsx
@@ -1,6 +1,19 @@
+import { setIsOpen } from '@/redux/features/profileModalSlice'
+import { RootState } from '@/redux/store'
 import React from 'react'
+import { useDispatch, useSelector } from 'react-redux'
 
-function LoginModal({ onClose }) {
+function LoginModal() {
+  const dispatch = useDispatch()
+
+  const isOpen = useSelector((state: RootState) => state.profile.isOpen)
+
+  function closeModal() {
+    if (isOpen) {
+      dispatch(setIsOpen(false))
+    }
+  }
+
   const LoginItem = [
     {
       title: '네이버로 로그인하기',
@@ -15,7 +28,7 @@ function LoginModal({ onClose }) {
     <div className='fixed inset-0 z-50 flex items-center justify-center bg-gray-500 bg-opacity-25'>
       <div className='w-[500px] rounded-lg bg-white'>
         <header className='flex w-full items-center justify-between border-b border-solid border-gray-200 p-4 pb-5'>
-          <button onClick={onClose} className='text-red-500'>
+          <button onClick={closeModal} className='text-red-500'>
             X
           </button>
           <div className='flex-grow text-center font-bold'>로그인 또는 회원가입</div>
